Create a database for storing data about sales using the Code First approach. The database should have the following tables:     • product (id, name, quantity, price)     • customer (id, name, email, credit_card_number)     • store_location (id, location_name)     • sale (id, product_id, customer_id, store_location_id, date) The relationships between the tables are as follows:     • Sale has one product and a product can be sold in many sales     • Sale has one customer and a customer can participate in many sales     • Sale has one store location and one store location can have many sales Hint You can use the following format to design your model classes:     • Product         ◦ int id         ◦ String name         ◦ Double quantity         ◦ BigDecimal price         ◦ Set<Sale> sales     • Customer         ◦ int id         ◦ String name         ◦ String email         ◦ String creditCardNumber         ◦ Set<Sale> sales     • StoreLocation         ◦ int id         ◦ String locationName         ◦ Set<Sale> sales     • Sale         ◦ int id         ◦ Product product         ◦ Customer customer         ◦ StoreLocation storeLocation         ◦ Date date

+++ ExerciseHibernateCodeFirst/src/main/java/A01/WizardDeposits.java
@@ -1,6 +1,7 @@
 package A01;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -37,13 +38,13 @@ public class WizardDeposits {
     private LocalDateTime depositStartDate;
 
     @Column(name = "deposit_amount", nullable = false)
-    private float depositAmount;
+    private BigDecimal depositAmount;
 
     @Column(name = "deposit_interest",nullable = false)
-    private float depositInterest;
+    private BigDecimal depositInterest;
 
     @Column(name = "deposit_charge",nullable = false)
-    private float depositCharge;
+    private BigDecimal depositCharge;
 
     @Column(name = "deposit_expiration_date", nullable = false)
     private LocalDateTime depositExpirationDate;
@@ -128,27 +129,27 @@ public class WizardDeposits {
         this.depositStartDate = depositStartDate;
     }
 
-    public float getDepositAmount() {
+    public BigDecimal getDepositAmount() {
         return depositAmount;
     }
 
-    public void setDepositAmount(float depositAmount) {
+    public void setDepositAmount(BigDecimal depositAmount) {
         this.depositAmount = depositAmount;
     }
 
-    public float getDepositInterest() {
+    public BigDecimal getDepositInterest() {
         return depositInterest;
     }
 
-    public void setDepositInterest(float depositInterest) {
+    public void setDepositInterest(BigDecimal depositInterest) {
         this.depositInterest = depositInterest;
     }
 
-    public float getDepositCharge() {
+    public BigDecimal getDepositCharge() {
         return depositCharge;
     }
 
-    public void setDepositCharge(float depositCharge) {
+    public void setDepositCharge(BigDecimal depositCharge) {
         this.depositCharge = depositCharge;
     }
